requirements.txt
selenium
beautifulsoup4
requests
pillow